move `isActive` and `disabled` props into `CommandItem`, and make some props optional
